requirements.txt
# Python wrapper around the SensorPush Bluetooth API
sensorpush-bleak >= 1.0.7